Remove Relational.Design package
