+++ src/EFCore.Design/Infrastructure/IPluralizer.cs
