+++ src/EFCore.Design/Scaffolding/Internal/CSharpNamer.cs
