+++ src/EFCore.Design/Scaffolding/Internal/CSharpUniqueNamer.cs
