+++ src/EFCore.Design/Scaffolding/Internal/CSharpUtilities.cs
