+++ src/EFCore.Design/Scaffolding/Internal/CandidateNamingService.cs
