+++ src/EFCore.Design/Scaffolding/Internal/IScaffoldingModelFactory.cs
