+++ src/EFCore.Design/Scaffolding/Internal/NullPluralizer.cs
