+++ src/EFCore.Design/Scaffolding/Internal/RelationalScaffoldingModelFactory.cs
@@ -256,7 +256,7 @@ namespace Microsoft.EntityFrameworkCore.Scaffolding
 
             if (keyBuilder == null)
             {
-                var errorMessage = RelationalDesignStrings.LogUnableToGenerateEntityType.GenerateMessage(table.DisplayName);
+                var errorMessage = DesignStrings.LogUnableToGenerateEntityType.GenerateMessage(table.DisplayName);
                 Logger.UnableToGenerateEntityTypeWarning(table.DisplayName);
 
                 var model = modelBuilder.Model;
