--- src/EFCore.Relational.Design/Properties/RelationalDesignStrings.Designer.tt
@@ -1,4 +1,0 @@
-<#
-    Session["ResourceFile"] = "RelationalDesignStrings.resx";
-#>
-<#@ include file="..\..\..\tools\Resources.tt" #>
