--- src/EFCore.Relational.Design/breakingchanges.netcore.json
@@ -1,73 +1,0 @@
-[
-  {
-    "TypeId": "public enum Microsoft.EntityFrameworkCore.Infrastructure.RelationalDesignEventId",
-    "Kind": "Removal"
-  },
-  {
-    "TypeId": "public class Microsoft.EntityFrameworkCore.Scaffolding.RelationalScaffoldingModelFactory : Microsoft.EntityFrameworkCore.Scaffolding.IScaffoldingModelFactory",
-    "MemberId": "protected virtual Microsoft.Extensions.Logging.ILogger get_Logger()",
-    "Kind": "Removal"
-  },
-  {
-    "TypeId": "public class Microsoft.EntityFrameworkCore.Scaffolding.RelationalScaffoldingModelFactory : Microsoft.EntityFrameworkCore.Scaffolding.IScaffoldingModelFactory",
-    "MemberId": "public .ctor(Microsoft.Extensions.Logging.ILoggerFactory loggerFactory, Microsoft.EntityFrameworkCore.Storage.IRelationalTypeMapper typeMapper, Microsoft.EntityFrameworkCore.Scaffolding.IDatabaseModelFactory databaseModelFactory, Microsoft.EntityFrameworkCore.Scaffolding.Internal.CandidateNamingService candidateNamingService)",
-    "Kind": "Removal"
-  },
-  {
-    "TypeId": "public class Microsoft.EntityFrameworkCore.Scaffolding.Metadata.ColumnModel : Microsoft.EntityFrameworkCore.Infrastructure.Annotatable",
-    "Kind": "Removal"
-  },
-  {
-    "TypeId": "public class Microsoft.EntityFrameworkCore.Scaffolding.Metadata.DatabaseModel : Microsoft.EntityFrameworkCore.Infrastructure.Annotatable",
-    "Kind": "Removal"
-  },
-  {
-    "TypeId": "public class Microsoft.EntityFrameworkCore.Scaffolding.Metadata.ForeignKeyColumnModel : Microsoft.EntityFrameworkCore.Infrastructure.Annotatable",
-    "Kind": "Removal"
-  },
-  {
-    "TypeId": "public class Microsoft.EntityFrameworkCore.Scaffolding.Metadata.ForeignKeyModel : Microsoft.EntityFrameworkCore.Infrastructure.Annotatable",
-    "Kind": "Removal"
-  },
-  {
-    "TypeId": "public class Microsoft.EntityFrameworkCore.Scaffolding.Metadata.IndexColumnModel : Microsoft.EntityFrameworkCore.Infrastructure.Annotatable",
-    "Kind": "Removal"
-  },
-  {
-    "TypeId": "public class Microsoft.EntityFrameworkCore.Scaffolding.Metadata.IndexModel : Microsoft.EntityFrameworkCore.Infrastructure.Annotatable",
-    "Kind": "Removal"
-  },
-  {
-    "TypeId": "public class Microsoft.EntityFrameworkCore.Scaffolding.Metadata.ScaffoldingModelAnnotations : Microsoft.EntityFrameworkCore.Metadata.RelationalModelAnnotations",
-    "Kind": "Removal"
-  },
-  {
-    "TypeId": "public class Microsoft.EntityFrameworkCore.Scaffolding.Metadata.ScaffoldingPropertyAnnotations : Microsoft.EntityFrameworkCore.Metadata.RelationalPropertyAnnotations",
-    "Kind": "Removal"
-  },
-  {
-    "TypeId": "public class Microsoft.EntityFrameworkCore.Scaffolding.Metadata.SequenceModel : Microsoft.EntityFrameworkCore.Infrastructure.Annotatable",
-    "Kind": "Removal"
-  },
-  {
-    "TypeId": "public class Microsoft.EntityFrameworkCore.Scaffolding.Metadata.TableModel : Microsoft.EntityFrameworkCore.Infrastructure.Annotatable",
-    "Kind": "Removal"
-  },
-  {
-    "TypeId": "public class Microsoft.EntityFrameworkCore.Scaffolding.ScaffoldingTypeMapper",
-    "Kind": "Removal"
-  },
-  {
-    "TypeId": "public class Microsoft.EntityFrameworkCore.Scaffolding.TypeScaffoldingInfo",
-    "Kind": "Removal"
-  },
-  {
-    "TypeId": "public static class Microsoft.EntityFrameworkCore.Metadata.ScaffoldingMetadataExtensions",
-    "Kind": "Removal"
-  },
-  {
-    "TypeId": "public class Microsoft.EntityFrameworkCore.Scaffolding.RelationalScaffoldingModelFactory : Microsoft.EntityFrameworkCore.Scaffolding.IScaffoldingModelFactory",
-    "MemberId": "protected virtual Microsoft.EntityFrameworkCore.Storage.RelationalTypeMapping GetTypeMapping(Microsoft.EntityFrameworkCore.Scaffolding.Metadata.ColumnModel column)",
-    "Kind": "Removal"
-  }
-]
