+++ src/EFCore.Relational/Diagnostics/RelationalEventId.cs
@@ -62,7 +62,32 @@ namespace Microsoft.EntityFrameworkCore.Diagnostics
 
             // Model validation events
             ModelValidationKeyDefaultValueWarning = CoreEventId.RelationalBaseId + 600,
-            BoolWithDefaultWarning
+            BoolWithDefaultWarning,
+
+            // Scaffolding warning events
+            MissingTableWarning = CoreEventId.CoreDesignBaseId + 700,
+            SequenceNotNamedWarning,
+            IndexColumnsNotMappedWarning,
+            ForeignKeyReferencesMissingTableWarning,
+            ForeignKeyReferencesMissingPrincipalTableWarning,
+            ForeignKeyColumnsNotMappedWarning,
+            ForeignKeyNotNamedWarning,
+            ForeignKeyColumnMissingWarning,
+            ForeignKeyColumnNotNamedWarning,
+            ForeignKeyPrincipalColumnMissingWarning,
+            ColumnNotNamedWarning,
+            IndexNotNamedWarning,
+            IndexTableMissingWarning,
+            IndexColumnNotNamedWarning,
+
+            // Scaffolding events
+            TableFound = CoreEventId.CoreDesignBaseId + 800,
+            TableSkipped,
+            ColumnSkipped,
+            IndexFound,
+            IndexColumnFound,
+            IndexColumnSkipped,
+            SequenceFound
         }
 
         private static readonly string _connectionPrefix = DbLoggerCategory.Database.Connection.Name + ".";
@@ -434,5 +459,134 @@ namespace Microsoft.EntityFrameworkCore.Diagnostics
         ///     </para>
         /// </summary>
         public static readonly EventId BoolWithDefaultWarning = MakeValidationId(Id.BoolWithDefaultWarning);
+
+        private static readonly string _scaffoldingPrefix = DbLoggerCategory.Scaffolding.Name + ".";
+        private static EventId MakeScaffoldingId(Id id) => new EventId((int)id, _scaffoldingPrefix + id);
+
+        /// <summary>
+        ///     The database is missing a table.
+        ///     This event is in the <see cref="DbLoggerCategory.Scaffolding" /> category.
+        /// </summary>
+        public static readonly EventId MissingTableWarning = MakeScaffoldingId(Id.MissingTableWarning);
+
+        /// <summary>
+        ///     The database has an unnamed sequence.
+        ///     This event is in the <see cref="DbLoggerCategory.Scaffolding" /> category.
+        /// </summary>
+        public static readonly EventId SequenceNotNamedWarning = MakeScaffoldingId(Id.SequenceNotNamedWarning);
+
+        /// <summary>
+        ///     Columns in an index were not mapped.
+        ///     This event is in the <see cref="DbLoggerCategory.Scaffolding" /> category.
+        /// </summary>
+        public static readonly EventId IndexColumnsNotMappedWarning = MakeScaffoldingId(Id.IndexColumnsNotMappedWarning);
+
+        /// <summary>
+        ///     A foreign key references a missing table.
+        ///     This event is in the <see cref="DbLoggerCategory.Scaffolding" /> category.
+        /// </summary>
+        public static readonly EventId ForeignKeyReferencesMissingTableWarning = MakeScaffoldingId(Id.ForeignKeyReferencesMissingTableWarning);
+
+        /// <summary>
+        ///     A foreign key references a missing table at the principal end.
+        ///     This event is in the <see cref="DbLoggerCategory.Scaffolding" /> category.
+        /// </summary>
+        public static readonly EventId ForeignKeyReferencesMissingPrincipalTableWarning = MakeScaffoldingId(Id.ForeignKeyReferencesMissingPrincipalTableWarning);
+
+        /// <summary>
+        ///     Columns in a foreign key were not mapped.
+        ///     This event is in the <see cref="DbLoggerCategory.Scaffolding" /> category.
+        /// </summary>
+        public static readonly EventId ForeignKeyColumnsNotMappedWarning = MakeScaffoldingId(Id.ForeignKeyColumnsNotMappedWarning);
+
+        /// <summary>
+        ///     A foreign key is not named.
+        ///     This event is in the <see cref="DbLoggerCategory.Scaffolding" /> category.
+        /// </summary>
+        public static readonly EventId ForeignKeyNotNamedWarning = MakeScaffoldingId(Id.ForeignKeyNotNamedWarning);
+
+        /// <summary>
+        ///     A foreign key column was not found.
+        ///     This event is in the <see cref="DbLoggerCategory.Scaffolding" /> category.
+        /// </summary>
+        public static readonly EventId ForeignKeyColumnMissingWarning = MakeScaffoldingId(Id.ForeignKeyColumnMissingWarning);
+
+        /// <summary>
+        ///     A column referenced by a foreign key constraint was not found.
+        ///     This event is in the <see cref="DbLoggerCategory.Scaffolding" /> category.
+        /// </summary>
+        public static readonly EventId ForeignKeyPrincipalColumnMissingWarning = MakeScaffoldingId(Id.ForeignKeyPrincipalColumnMissingWarning);
+
+        /// <summary>
+        ///     A foreign key column was not named.
+        ///     This event is in the <see cref="DbLoggerCategory.Scaffolding" /> category.
+        /// </summary>
+        public static readonly EventId ForeignKeyColumnNotNamedWarning = MakeScaffoldingId(Id.ForeignKeyColumnNotNamedWarning);
+
+        /// <summary>
+        ///     A column is not named.
+        ///     This event is in the <see cref="DbLoggerCategory.Scaffolding" /> category.
+        /// </summary>
+        public static readonly EventId ColumnNotNamedWarning = MakeScaffoldingId(Id.ColumnNotNamedWarning);
+
+        /// <summary>
+        ///     An index is not named.
+        ///     This event is in the <see cref="DbLoggerCategory.Scaffolding" /> category.
+        /// </summary>
+        public static readonly EventId IndexNotNamedWarning = MakeScaffoldingId(Id.IndexNotNamedWarning);
+
+        /// <summary>
+        ///     The table referened by an index was not found.
+        ///     This event is in the <see cref="DbLoggerCategory.Scaffolding" /> category.
+        /// </summary>
+        public static readonly EventId IndexTableMissingWarning = MakeScaffoldingId(Id.IndexTableMissingWarning);
+
+        /// <summary>
+        ///     An index column was not named.
+        ///     This event is in the <see cref="DbLoggerCategory.Scaffolding" /> category.
+        /// </summary>
+        public static readonly EventId IndexColumnNotNamedWarning = MakeScaffoldingId(Id.IndexColumnNotNamedWarning);
+
+        /// <summary>
+        ///     A table was found.
+        ///     This event is in the <see cref="DbLoggerCategory.Scaffolding" /> category.
+        /// </summary>
+        public static readonly EventId TableFound = MakeScaffoldingId(Id.TableFound);
+
+        /// <summary>
+        ///     A table was skipped.
+        ///     This event is in the <see cref="DbLoggerCategory.Scaffolding" /> category.
+        /// </summary>
+        public static readonly EventId TableSkipped = MakeScaffoldingId(Id.TableSkipped);
+
+        /// <summary>
+        ///     A column was skipped.
+        ///     This event is in the <see cref="DbLoggerCategory.Scaffolding" /> category.
+        /// </summary>
+        public static readonly EventId ColumnSkipped = MakeScaffoldingId(Id.ColumnSkipped);
+
+        /// <summary>
+        ///     An index was found.
+        ///     This event is in the <see cref="DbLoggerCategory.Scaffolding" /> category.
+        /// </summary>
+        public static readonly EventId IndexFound = MakeScaffoldingId(Id.IndexFound);
+
+        /// <summary>
+        ///     An index was skipped.
+        ///     This event is in the <see cref="DbLoggerCategory.Scaffolding" /> category.
+        /// </summary>
+        public static readonly EventId IndexColumnFound = MakeScaffoldingId(Id.IndexColumnFound);
+
+        /// <summary>
+        ///     A column of an index was skipped.
+        ///     This event is in the <see cref="DbLoggerCategory.Scaffolding" /> category.
+        /// </summary>
+        public static readonly EventId IndexColumnSkipped = MakeScaffoldingId(Id.IndexColumnSkipped);
+
+        /// <summary>
+        ///     A sequence was found.
+        ///     This event is in the <see cref="DbLoggerCategory.Scaffolding" /> category.
+        /// </summary>
+        public static readonly EventId SequenceFound = MakeScaffoldingId(Id.SequenceFound);
     }
 }
