+++ test/EFCore.Design.Tests/Scaffolding/Internal/CSharpNamerTest.cs
