+++ test/EFCore.Design.Tests/Scaffolding/Internal/CSharpUniqueNamerTest.cs
