+++ test/EFCore.Design.Tests/Scaffolding/Internal/CandidateNamingServiceTest.cs
