+++ test/EFCore.Design.Tests/Scaffolding/Internal/NullPluralizerTest.cs
