+++ test/EFCore.Design.Tests/Scaffolding/Internal/ScaffoldingMetadataExtenstionsTest.cs
