--- test/EFCore.Relational.Design.Tests/ApiConsistencyTest.cs
@@ -1,14 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System.Reflection;
-using Microsoft.EntityFrameworkCore.Scaffolding;
-
-namespace Microsoft.EntityFrameworkCore
-{
-    public class ApiConsistencyTest : ApiConsistencyTestBase
-    {
-        protected override Assembly TargetAssembly
-            => typeof(IScaffoldingModelFactory).GetTypeInfo().Assembly;
-    }
-}
