+++ test/ef.Tests/AppDomainOperationExecutorTest.cs
@@ -54,7 +54,6 @@ namespace Microsoft.EntityFrameworkCore.Tools
                         BuildReference.ByName("Microsoft.EntityFrameworkCore", true),
                         BuildReference.ByName("Microsoft.EntityFrameworkCore.Design", true),
                         BuildReference.ByName("Microsoft.EntityFrameworkCore.Relational", true),
-                        BuildReference.ByName("Microsoft.EntityFrameworkCore.Relational.Design", true),
                         BuildReference.ByName("Microsoft.EntityFrameworkCore.SqlServer", true),
                         BuildReference.ByName("Microsoft.Extensions.Caching.Abstractions", true),
                         BuildReference.ByName("Microsoft.Extensions.Caching.Memory", true),
@@ -160,7 +159,6 @@ namespace Microsoft.EntityFrameworkCore.Tools
                         BuildReference.ByName("Microsoft.EntityFrameworkCore"),
                         BuildReference.ByName("Microsoft.EntityFrameworkCore.Design", true),
                         BuildReference.ByName("Microsoft.EntityFrameworkCore.Relational", true),
-                        BuildReference.ByName("Microsoft.EntityFrameworkCore.Relational.Design", true),
                         BuildReference.ByName("Microsoft.Extensions.DependencyInjection", true),
                         BuildReference.ByName("Microsoft.Extensions.DependencyInjection.Abstractions", true),
                         BuildReference.ByName("Microsoft.Extensions.Logging", true),
@@ -232,7 +230,6 @@ namespace Microsoft.EntityFrameworkCore.Tools
                         BuildReference.ByName("Microsoft.EntityFrameworkCore", true),
                         BuildReference.ByName("Microsoft.EntityFrameworkCore.Design", true),
                         BuildReference.ByName("Microsoft.EntityFrameworkCore.Relational", true),
-                        BuildReference.ByName("Microsoft.EntityFrameworkCore.Relational.Design", true),
                         BuildReference.ByName("Microsoft.EntityFrameworkCore.SqlServer", true),
                         BuildReference.ByName("Microsoft.Extensions.Caching.Abstractions", true),
                         BuildReference.ByName("Microsoft.Extensions.Caching.Memory", true),
@@ -304,7 +301,6 @@ namespace Microsoft.EntityFrameworkCore.Tools
                         BuildReference.ByName("Microsoft.EntityFrameworkCore", true),
                         BuildReference.ByName("Microsoft.EntityFrameworkCore.Design", true),
                         BuildReference.ByName("Microsoft.EntityFrameworkCore.Relational", true),
-                        BuildReference.ByName("Microsoft.EntityFrameworkCore.Relational.Design", true),
                         BuildReference.ByName("Microsoft.Extensions.DependencyInjection", true),
                         BuildReference.ByName("Microsoft.Extensions.DependencyInjection.Abstractions", true),
                         BuildReference.ByName("Microsoft.Extensions.Logging", true),
